+++ src/clified/__init__.py
@@ -3,4 +3,4 @@
 from clified.logging import Logger
 
 __all__ = ["Logger", "__version__"]
-__version__ = "0.7.2"
+__version__ = "0.7.3"